[fix] attack-structure
